+++ extractor/base/utils.py
@@ -1,7 +1,7 @@
 """TODO:"""
 
-# import json
-import os
+
+# TODO: can incorporate both of these into the extractor?
 
 
 def check_row_quant_safety(paged_list, range_start, range_end) -> int:
@@ -39,26 +39,3 @@ def clean_str(str_to_clean):
         output_str = output_str.replace("\n", "")
 
     return output_str.strip()
-
-
-def verify_dirs(file_path):
-    """
-    verifies that the path to parameter exists or creates that path
-
-    :param file_path str: file path to check or create
-    """
-
-    # get only the last item in the file path, i.e. the item after the last slash
-    # ( the file name )
-    stripped_path_list = file_path.rsplit("/", 1)
-
-    # determine if the split performed above created two separate items. If not
-    # ( meaning the list length is 1 ), only a file name was given and that file
-    # would be created in the same directory as the extractor. If the length is
-    # greater than 1, we will have to either create or verify the existence of
-    # the path to the file being created
-    if len(stripped_path_list) > 1:
-
-        path = stripped_path_list[0]
-
-        os.makedirs(path, exist_ok=True)